Fix linting circular dependencies
